Add a parser for parsing types without the + character
This allows us to correctly parse types in more situations. For example,
`&T` should not allow `+` characters in `T` without brackets, as `&`
binds more tightly than the `+` operator, which is now correctly
handled.

I don't believe that I correctly ignore `+` in all required places in
ty.rs yet, but it's a first step. This set of changes allows us to pass
all of our current parsing tests.

In addition, this allows us to fix our parsing of `as T` expressions in
the next patch.