Add helpful accessors to Generics
NOTE: This PR is not yet ready to be merged. This adds only one of the
three needed accessors, and has no tests. I wanted to make sure that
this was a feature that would be accepted before spending the time to
finish this.

Within Diesel we have a bunch of places that do things like
`generics.params.iter()`. It'd be very helpful for syn to provide these
iterators for us.

This PR as written only contains the iterator for `type_params` (holy
crap I hate that name can we please rename `TypeParam` to
`TypeParameter`). The other iterators will be mechanically
identical.

If you're good with this feature I will implement the remaining
iterators and add tests.