Add attrs to variadic in bare fn type
This breaks the TypeBareFn syntax tree but it's so obscure that I don't
expect this to affect any existing code. Even in code that touches
TypeBareFn, it would be unusual for it to be doing anything other than
calling .is_some() on variadic or assigning None to it. Let me know if I
got it wrong.